Fixing bug of interactions and first attempt of placeholders
