+++ Assets/Placeholder.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d53d0c83369b6b143a188354a7cabf56
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
