+++ Assets/Prefabs/Card.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f99572b5d92480447a2097dea2a74740
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
