+++ Assets/Prefabs/Table.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 99bfed509bf69df40bae44cf79b385d6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
