+++ Assets/Scripts/AnimateHandOnInput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98b6cee45036b81408ab08addfcc8ac5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
